apps/api/requirements.txt
# FastAPI & Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# Redis & Queue
redis==5.0.1
rq==1.16.1

# Auth
pyjwt[crypto]==2.8.0
cryptography==42.0.0

# LLM Providers
openai==1.10.0
google-generativeai==0.3.2
# Perplexity uses standard HTTP API (httpx already included)

# HTTP & Scraping
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Storage
boto3==1.34.24

# Payments
stripe==7.11.0

# Email
postmarker==1.0

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0
sentry-sdk[fastapi]==1.39.2

# Utilities
python-dotenv==1.0.0
python-slugify==8.0.1
python-dateutil==2.8.2
pydantic-extra-types==2.4.1
tenacity==8.2.3
email-validator==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0
httpx==0.26.0

# Linting & Formatting
black==24.1.1
isort==5.13.2
ruff==0.1.14
mypy==1.8.0
bandit==1.7.6

